add `--check-only` flag to `generate` command (fixes #85) (#89)
If present, this flag causes the generate command to resolve all dependencies as usual, but does not write the files.
Instead, it checks that each generated file on disk has the expected content.  If any files do not match the expected
content, they are logged and bazel-deps exits with code 2; bazel-deps exits 0 if everything matches.