Switch from java_library to scala_import
Switch from using java_library to scala_import for Scala dependencies.

Notes:

1.
java dependencies still use java_library; ideally we would switch to
java_import for consistency but java rules are still in flux so
we're sticking to what works

2.
the label passed to `jars` attribute might actually not point directly
at the :file target (which is a filegroup) but at the synthetic
java_import so we're technically not following the contract that says
`jars` should have labels pointing at jar files. However, with the current
implementation of scala_import that's allowed and it actually helps
IntelliJ to resolve the source artifact for a jar. I think this behavior
is actually a bug in the scala_import implementation.