Add support for scala_import (#101)
* Update scala rules and load scala_import

* Switch from java_library to scala_import

Switch from using java_library to scala_import for Scala dependencies.

Notes:

1.
java dependencies still use java_library; ideally we would switch to
java_import for consistency but java rules are still in flux so
we're sticking to what works

2.
the label passed to `jars` attribute might actually not point directly
at the :file target (which is a filegroup) but at the synthetic
java_import so we're technically not following the contract that says
`jars` should have labels pointing at jar files. However, with the current
implementation of scala_import that's allowed and it actually helps
IntelliJ to resolve the source artifact for a jar. I think this behavior
is actually a bug in the scala_import implementation.

* Check in all BUILD files switched to scala_import

The import into IntelliJ works!

* Load scala_import explicitly instead of using prelude

Add `load` for scala_import to buildHeader element and get rid of that load
from prelude. The scala_import is only used in 3rdparty so mentioning it
in dependencies.yaml file is enough.

We want to move away from using prelude_bazel.

* Update TODO comment on *_library -> *_import switch

Address review feedback