Fix SHA1 check
In Resolver.scala a SHA1 value was validated using:

  scala.math.BigInt(hexString,16).asByteArray.length == 20.

This does not work if the SHA1 value starts with a high order bit, i.e, 8-f.
In that case the byte array representation of the BigInt has an additional sign
byte and the length of the byte array is 21.

The fix is to switch from using BigInt for validation of SHA1 values to a
simple syntactic check on the string, i.e., each character is in [0-9A-Fa-f]
and the lengt of the string is 40.

This fixes #120.