Use SLF4J logging in MakeDeps and related code.
We were already including an SLF4J dependency, so all we had to do was
to start using it! This commit adds a Verbosity type, along with the
Decline machinery to be able to pass log levels as command-line
arguments.

In the future we should add verbosity to the other bazel-deps
commands, and replace all error/info reporting println's with logging
instead. (Some println's represent program output; these should
probably stay as-is.)